modify: Don't allow resizing of transition data
Prior to this change, ReplaceActionTransitions() only complained about a
change in size if a transition's size increased.  This error checking is
too relaxed: any change in size will cause a mismatch in WL.EXE's
hardcoded block offsets.

This commit tightens this error check.  Now if the size changes in
either direction the function fails.