msq: Rename Block to Body
This is a large compatibility-breaking change.

The basic problem was that `msq.Block` contained some fields that it
really shouldn't have: Offset and Hdr.

Modifications to one block in a GAME file could cause another block's
Offset field to become inaccurate.  This is obviously not good.

The Hdr field duplicated some state that is present in the EncSection
and PlainSection fields.  Namely, the Hdr.Xor fields are derived from
the block body.  If the body changes, the header becomes out of sync.
This problem can be avoided by not storing the header in a block.
Instead, we just calculate the header on the fly as needed.

The solution is to reduce the responsibility of the `msq.Block` type so
that it only contains the block body (`EncSection` and `PlainSection`).
This modification warrants a name change, so the type is now named
`msq.Body`.

A new type with the Hdr and Offset fields was created: `msq.Desc` (block
descriptor).